backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.1
requests>=2.31.0
aiosqlite>=0.19.0
# IA
openai>=1.12.0
google-genai>=0.3.0
# Documents
pypdf>=4.0.1
python-docx>=1.1.0
jinja2>=3.1.3
# Paiement
stripe>=8.1.0
# Websockets
websockets>=12.0
# Securite (Auth & JWT)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
# Base de donnees PostgreSQL
psycopg2-binary>=2.9.9
asyncpg>=0.29.0